requirements.txt
scipy
numpy

matplotlib
opencv-python
Pillow

open3d

gpytorch
transformers